webapp/loggedin folder: moved RecordReducer under webapp/store/ui/record
