+++ webapp/loggedin/surveyViews/reducer.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux'
 import * as RecordState from './record/recordState'
 import * as SurveyFormState from './surveyForm/surveyFormState'
 
-import record from './record/reducer'
+import record from '../../store/ui/record/reducer'
 import surveyForm from './surveyForm/reducer'
 
 export default combineReducers({
