+++ webapp/store/ui/record/index.js
@@ -1,0 +1,3 @@
+import RecordReducer from './reducer'
+
+export { RecordReducer }
